Added smarter color removal for the color wheel in colorblind settings
Similar colors are now stored in a list. Therefore, when removing a color from the texture, it is not needed to check the similarity between all pixels, but just between the few different color in the list.

This list is initialised once when in the main menu

Debug: remove residual color lines + avoid selection of alpha=0 colors