Automatic solver (#2)
* added level solver script

Added class graph, action and node

It is possible to get through a whole level with the solver, and rewind to the beginning

* Beginning of graph solver

* Graph solver v1.0

The solver is working and able to solve entirely the first level

It uses an already visited state memory to avoid repetitive or looping states

Stats 1rst level: Last node: 3947 / greatest node: 4065 / already visited nodes: 2544

* Bug correction on findActions step

The differentAction condition wronlgy removed some possibilities. In particular when at least one tube was used in a previous move

* Updated levels up to 49 with solver (43 not working though)

* Added version to level generator + level up to 55

It is now possible to customise level generator to fit different function in the color choice. It is done so it can be easier to setup complex levels

* Build v0.7.0

* Updated level generator to allow multiple versions

* Levels OK up to 71

* levels ok up to 77