Creation of power mechanics (#4)
* Added robot power template

* Updated gitignore

* Added color wheel perso

* Added Roll back one + is still winnable powers

* Robot power backend done for isWinnable and next winnable move powers

* Added delete one color power backend

* Updated delete one color bonus to reinitialise main solver

The reason to do it is to avoid rewinding a move that should not be possible because of the new colors

* Started token list management

* Added token check for a singular power

Back prog of power's token done

* Added token management and GUI

Tokens are represented as small indicators on top of the screen. Each token represent a color, and the number of token available for a given color

Initiatialisation, load, and update of the tokens in player prefs is functional

* Added token sprites

* Added power buttons

The state machine is updated to incorporate the state when the player is picking and using a power

An additional screen is used to select the powers. The screen is available when the robot is clicked by the player.

* Linked power usage with token availability

Power will not be available (buttons not interactable) if the player does not have enough tokens

When used, the powers consums corresponding tokens