Ads manager done
Ads can be loaded and managed through an ad manager GameObject. The script that handle when an ad is shown is the ad_manager script

Currently, an ad banner is shown during level selection, and a vidoe is displayed when loading (or re-loading) a level

TODO: Maybe  add probability of display ad video when launching level. Add reward ad to increases possessed tokens