Advertisement added to the game (#6)
* Started adding advertisements

Initialisation of a first ad on the screen OK

* Ads manager done

Ads can be loaded and managed through an ad manager GameObject. The script that handle when an ad is shown is the ad_manager script

Currently, an ad banner is shown during level selection, and a vidoe is displayed when loading (or re-loading) a level

TODO: Maybe  add probability of display ad video when launching level. Add reward ad to increases possessed tokens

* Ads setup for each scene DONE

Ads are now fully integrated in the game

Banners are present in level Selection

video ad is shown with a defined probability (40% for now) when loading level

Reward ad available to increase tokens possession. Button to launch ad is on the token icon